requirements/test.txt
# Test dependencies go here.
-r base.txt



coverage==4.2
django_coverage_plugin==1.3
flake8==3.2.0 # pyup: != 2.6.0
django-test-plus==1.0.16
factory_boy==2.7.0

# pytest
pytest-django==3.0.0
pytest-sugar==0.7.1
